Add files via upload
This is the welcome worksheet that I gave the students while they were filing into the classroom on the first day, and the presentation that I used to open the day. I would ask the group who felt confident about their answer about each question, and then show the slide with the answer, finally tying it all together with the fact that what has been valuable in the past, became common as society advanced, and it's important not to assume we know today what will hold its value into the future, because something new could come along and disrupt those assumptions!